@@ -7,29 +7,29 @@
         "@types/node": "^12.0.0",
         "@types/react": "^16.9.0",
         "@types/react-dom": "^16.9.0",
-        "@unform/core": "^2.1.3",
-        "@unform/web": "^2.1.3",
+        "@unform/core": "^2.1.6",
+        "@unform/web": "^2.1.6",
         "axios": "^0.21.2",
+        "bulma": "^0.9.3",
         "date-fns": "^2.16.1",
         "polished": "^3.6.5",
-        "react": "^16.13.1",
-        "react-day-picker": "^7.4.8",
-        "react-dom": "^16.13.1",
+        "react": "17.0.2",
+        "react-bulma-components": "^4.1.0",
+        "react-day-picker": "^7.4.10",
+        "react-dom": "17.0.2",
         "react-icons": "^3.10.0",
-        "react-router-dom": "^5.2.0",
-        "react-scripts": "3.4.1",
+        "react-router-dom": "^6.2",
         "react-spring": "^8.0.27",
-        "styled-components": "^5.1.1",
-        "typescript": "~3.8.0",
+        "styled-components": "^5.3",
+        "typescript": "^4.5.4",
         "uuidv4": "^6.1.1",
-        "yup": "^0.29.1"
+        "yup": "^0.29.1",
+        "vite": "^2.8.0",
+        "@vitejs/plugin-react": "^1.1.1"
     },
     "scripts": {
-        "start": "react-scripts start",
-        "build": "react-scripts build",
-        "test": "react-scripts test",
-        "test:coverage": "react-scripts test --coverage --watchAll false",
-        "eject": "react-scripts eject"
+        "start": "vite",
+        "build": "vite build"
     },
     "eslintConfig": {
         "extends": "react-app"
@@ -58,11 +58,11 @@
         "source.fixAll.eslint": true
     },
     "devDependencies": {
+        "@testing-library/dom": "^7.24.3",
         "@testing-library/jest-dom": "^4.2.4",
         "@testing-library/react": "^11.0.4",
         "@testing-library/user-event": "^7.1.2",
         "@types/jest": "^24.0.0",
-        "@testing-library/dom": "^7.24.3",
         "@types/react-router-dom": "^5.1.5",
         "@types/styled-components": "^5.1.0",
         "@types/yup": "^0.29.3",
