@@ -3,7 +3,7 @@ import { FiArrowLeft, FiUser, FiLock, FiMail } from "react-icons/fi";
 import { FormHandles } from "@unform/core";
 import { Form } from "@unform/web";
 import * as Yup from "yup";
-import { Link, useHistory } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 import api from "../../services/api";
 import { useToast } from "../../hooks/toast";
 import getValidationErrors from "../../utils/getValidationErrors";
@@ -24,52 +24,50 @@ interface SignUpFormData {
 const SignUp: React.FC = () => {
     const formRef = useRef<FormHandles>(null);
     const { addToast } = useToast();
-    const history = useHistory();
+    const navigate = useNavigate();
 
-    const handleSubmit = useCallback(async (data: SignUpFormData) => {
-        try {
-            formRef.current?.setErrors({});
-            const schema = Yup.object().shape({
-                name: Yup.string().required("Nome é obrigatório"),
-                email: Yup.string()
-                    .required("E-mail obrigatório")
-                    .email("Digite um e-mail válido"),
-                password: Yup.string().min(6, "no minimo 6 dígitos"),
-            });
+    const handleSubmit = useCallback(
+        async (data: SignUpFormData) => {
+            try {
+                formRef.current?.setErrors({});
+                const schema = Yup.object().shape({
+                    name: Yup.string().required("Nome é obrigatório"),
+                    email: Yup.string()
+                        .required("E-mail obrigatório")
+                        .email("Digite um e-mail válido"),
+                    password: Yup.string().min(6, "no minimo 6 dígitos"),
+                });
 
-            await schema.validate(data, {
-                abortEarly: false,
-            });
+                await schema.validate(data, {
+                    abortEarly: false,
+                });
 
-            await api.post("/users", data);
+                await api.post("/users", data);
 
-            history.push("/");
+                navigate("/");
 
-            addToast({
-                type: "success",
-                title: "Cadastro realizado",
-                description: "Pode fazer seu login no Gobarber",
-            });
-        } catch (err) {
-            if (err instanceof Yup.ValidationError) {
-                const errors = getValidationErrors(err);
-                formRef.current?.setErrors(errors);
-            } else {
-                console.log(err);
-                return;
+                addToast({
+                    type: "success",
+                    title: "Cadastro realizado",
+                    description: "Pode fazer seu login no Gobarber",
+                });
+            } catch (err) {
+                if (err instanceof Yup.ValidationError) {
+                    const errors = getValidationErrors(err);
+                    formRef.current?.setErrors(errors);
+                } else {
+                    console.log(err);
+                    return;
+                }
+
+                addToast({
+                    type: "error",
+                    title: "erro na autenticação",
+                    description: "ocorreu um erro ao logar",
+                });
             }
-
-            addToast({
-                type: "error",
-                title: "erro na autenticação",
-                description: "ocorreu um erro ao logar",
-            });
-        }
-
         },
-
-
-        [addToast, history]
+        [addToast, navigate]
     );
 
     return (
@@ -99,8 +97,8 @@ const SignUp: React.FC = () => {
                     </Form>
                     <Link to="/">
                         <FiArrowLeft />
-                            Voltar para login
-                        </Link>
+                        Voltar para login
+                    </Link>
                 </AnimationContainer>
             </Content>
         </Container>
