@@ -1,11 +1,15 @@
+/* eslint-disable import/no-duplicates */
 import React, { useState, useCallback, useEffect, useMemo } from "react";
-import {Link} from 'react-router-dom';
-import { isToday, format, parseISO, isAfter } from 'date-fns';
-import ptBR from 'date-fns/locale/pt-BR';
-import DayPicker, { DayModifiers } from 'react-day-picker';
-import 'react-day-picker/lib/style.css';
+import { Link, useNavigate } from "react-router-dom";
+import { isToday, format, parseISO, isAfter } from "date-fns";
+import ptBR from "date-fns/locale/pt-BR";
 
+import DayPicker, { DayModifiers } from "react-day-picker";
+import "react-day-picker/lib/style.css";
 
+import { Button } from "react-bulma-components";
+
+import { FiPower, FiClock, FiPlus } from "react-icons/fi";
 import {
     Container,
     Header,
@@ -17,12 +21,12 @@ import {
     NextAppointment,
     Appointment,
     Section,
-} from './styles';
+} from "./styles";
 
-import logoImg from '../../assets/logo-white.png';
-import { FiPower, FiClock } from "react-icons/fi";
+import "bulma/css/bulma.min.css";
+import logoImg from "../../assets/logo-white.png";
 import { useAuth } from "../../hooks/auth";
-import api from '../../services/api';
+import api from "../../services/api";
 
 interface MonthAvailabilityItem {
     day: number;
@@ -39,57 +43,61 @@ interface Appointment {
     };
 }
 
-
 interface CalendarModifiers extends DayModifiers {
     available: boolean;
 }
 
-
 const Dashboard: React.FC = () => {
-
     const { user, signOut } = useAuth();
+    const navigate = useNavigate();
     const [selectedDate, setSelectedDate] = useState(new Date());
     const [currentMonth, setCurrentMonth] = useState(new Date());
-    const [monthAvailability, setMonthAvailability] = useState<MonthAvailabilityItem[]>([]);
+    const [monthAvailability, setMonthAvailability] = useState<
+        MonthAvailabilityItem[]
+    >([]);
 
     const [appointments, setAppointments] = useState<Appointment[]>([]);
 
-    const handleDateChange = useCallback((day: Date, modifiers: DayModifiers) => {
-
-        if (modifiers.available && !modifiers.disabled) {
-            setSelectedDate(day);
-        }
-
-    }, []);
+    const handleDateChange = useCallback(
+        (day: Date, modifiers: DayModifiers) => {
+            if (modifiers.available && !modifiers.disabled) {
+                setSelectedDate(day);
+            }
+        },
+        []
+    );
 
     const handleMonthChange = useCallback((month: Date) => {
         setCurrentMonth(month);
     }, []);
+
+    const handleNewAppointment = useCallback(async () => {
+        navigate(`/appointments/new?date=${selectedDate.getTime()}`);
+    }, [navigate, selectedDate]);
 
     useEffect(() => {
         api.get(`/providers/${user.id}/month-availability`, {
             params: {
                 year: currentMonth.getFullYear(),
                 month: currentMonth.getMonth() + 1,
-            }
-        }).then(response => {
+            },
+        }).then((response) => {
             setMonthAvailability(response.data);
         });
-
-    }, [currentMonth, user.id])
+    }, [currentMonth, user.id]);
 
     useEffect(() => {
-        api.get<Appointment[]>('/appointments/me', {
+        api.get<Appointment[]>("/appointments/me", {
             params: {
                 year: selectedDate.getFullYear(),
                 month: selectedDate.getMonth() + 1,
                 day: selectedDate.getDate(),
-            }
-        }).then(response => {
-            const appointmentsFormatted = response.data.map(appointment => {
+            },
+        }).then((response) => {
+            const appointmentsFormatted = response.data.map((appointment) => {
                 return {
                     ...appointment,
-                    hourFormatted: format(parseISO(appointment.date), 'HH:mm'),
+                    hourFormatted: format(parseISO(appointment.date), "HH:mm"),
                 };
             });
             setAppointments(appointmentsFormatted);
@@ -98,8 +106,8 @@ const Dashboard: React.FC = () => {
 
     const disabledDays = useMemo(() => {
         const dates = monthAvailability
-            .filter(monthDay => monthDay.available === false)
-            .map(monthDay => {
+            .filter((monthDay) => monthDay.available === false)
+            .map((monthDay) => {
                 const year = currentMonth.getFullYear();
                 const month = currentMonth.getMonth();
 
@@ -107,9 +115,7 @@ const Dashboard: React.FC = () => {
             });
 
         return dates;
-
     }, [currentMonth, monthAvailability]);
-
 
     const selectedDateAsText = useMemo(() => {
         return format(selectedDate, "'Dia' dd 'de' MMMM", {
@@ -118,41 +124,37 @@ const Dashboard: React.FC = () => {
     }, [selectedDate]);
 
     const selectedWeekDay = useMemo(() => {
-        return format(selectedDate, 'cccc', {
+        return format(selectedDate, "cccc", {
             locale: ptBR,
-        })
+        });
     }, [selectedDate]);
 
     const morningAppointments = useMemo(() => {
-        return appointments.filter(appointment => {
+        return appointments.filter((appointment) => {
             return parseISO(appointment.date).getHours() < 12;
         });
-
     }, [appointments]);
 
     const afternoonAppointments = useMemo(() => {
-        return appointments.filter(appointment => {
+        return appointments.filter((appointment) => {
             return parseISO(appointment.date).getHours() >= 12;
         });
-
     }, [appointments]);
 
     const nextAppointment = useMemo(() => {
-        return appointments.find(appointment =>
-            isAfter(parseISO(appointment.date), new Date()),
+        return appointments.find((appointment) =>
+            isAfter(parseISO(appointment.date), new Date())
         );
     }, [appointments]);
 
     return (
-
         <Container>
             <Header>
                 <HeaderContent>
                     <img src={logoImg} alt="Maniclub" />
 
                     <Profile>
-                        <img
-                            src={user.avatar_url} alt={user.name} />
+                        <img src={user.avatar_url} alt={user.name} />
 
                         <div>
                             <span>Bem vindo!</span>
@@ -162,7 +164,7 @@ const Dashboard: React.FC = () => {
                         </div>
                     </Profile>
 
-                    <button type="button" onClick={signOut} >
+                    <button type="button" onClick={signOut}>
                         <FiPower />
                     </button>
                 </HeaderContent>
@@ -179,18 +181,19 @@ const Dashboard: React.FC = () => {
 
                     {isToday(selectedDate) && nextAppointment && (
                         <NextAppointment>
-                        <strong>Agendamento a seguir</strong>
-                        <div>
-                            <img src={nextAppointment.user.avatar_url}
-                                alt={nextAppointment.user.name} />
-                            <strong>{nextAppointment.user.name}</strong>
-                            <span>
-                                <FiClock />
-                                 {nextAppointment.hourFormatted}
-                            </span>
-                        </div>
-
-                    </NextAppointment>
+                            <strong>Agendamento a seguir</strong>
+                            <div>
+                                <img
+                                    src={nextAppointment.user.avatar_url}
+                                    alt={nextAppointment.user.name}
+                                />
+                                <strong>{nextAppointment.user.name}</strong>
+                                <span>
+                                    <FiClock />
+                                    {nextAppointment.hourFormatted}
+                                </span>
+                            </div>
+                        </NextAppointment>
                     )}
 
                     <Section>
@@ -200,25 +203,22 @@ const Dashboard: React.FC = () => {
                             <p>Nenhum agendamento neste período !!</p>
                         )}
 
-                        {morningAppointments.map(appointment => (
-
+                        {morningAppointments.map((appointment) => (
                             <Appointment key={appointment.id}>
-
                                 <span>
                                     <FiClock />
                                     {appointment.hourFormatted}
                                 </span>
 
                                 <div>
-                                    <img src={appointment.user.avatar_url}
-                                        alt={appointment.user.name} />
+                                    <img
+                                        src={appointment.user.avatar_url}
+                                        alt={appointment.user.name}
+                                    />
                                     <strong>{appointment.user.name}</strong>
                                 </div>
-
                             </Appointment>
-
                         ))}
-
                     </Section>
 
                     <Section>
@@ -228,59 +228,54 @@ const Dashboard: React.FC = () => {
                             <p>Nenhum agendamento neste período !!</p>
                         )}
 
-                        {afternoonAppointments.map(appointment => (
-
+                        {afternoonAppointments.map((appointment) => (
                             <Appointment key={appointment.id}>
-
                                 <span>
                                     <FiClock />
                                     {appointment.hourFormatted}
                                 </span>
 
                                 <div>
-                                    <img src={appointment.user.avatar_url}
-                                        alt={appointment.user.name} />
+                                    <img
+                                        src={appointment.user.avatar_url}
+                                        alt={appointment.user.name}
+                                    />
                                     <strong>{appointment.user.name}</strong>
                                 </div>
-
                             </Appointment>
-
                         ))}
                     </Section>
                 </Schedule>
                 <Calendar>
                     <DayPicker
-                        weekdaysShort={['D', 'S', 'T', 'Q', 'Q', 'S', 'S']}
+                        weekdaysShort={["D", "S", "T", "Q", "Q", "S", "S"]}
                         fromMonth={new Date()}
                         disabledDays={[{ daysOfWeek: [0, 6] }, ...disabledDays]}
                         modifiers={{
-                            available: { daysOfWeek: [1, 2, 3, 4, 5] }
+                            available: { daysOfWeek: [1, 2, 3, 4, 5] },
                         }}
                         onMonthChange={handleMonthChange}
                         selectedDays={selectedDate}
                         onDayClick={handleDateChange}
                         months={[
-                            'Janeiro',
-                            'Fevereiro',
-                            'Março',
-                            'Abril',
-                            'Maio',
-                            'Junho',
-                            'Julho',
-                            'Agosto',
-                            'Setembro',
-                            'Outubro',
-                            'Novembro',
-                            'Dezembro',
+                            "Janeiro",
+                            "Fevereiro",
+                            "Março",
+                            "Abril",
+                            "Maio",
+                            "Junho",
+                            "Julho",
+                            "Agosto",
+                            "Setembro",
+                            "Outubro",
+                            "Novembro",
+                            "Dezembro",
                         ]}
                     />
                 </Calendar>
             </Content>
         </Container>
-
     );
-}
-
-
+};
 
 export default Dashboard;
