@@ -1,7 +1,7 @@
 import axios from "axios";
 
 const api = axios.create({
-    baseURL: process.env.REACT_APP_URL,
+    baseURL: import.meta.env.VITE_REACT_APP_URL,
 });
 
 export default api;
